+++ osdu/client/aws.py
@@ -20,7 +20,7 @@ class AwsOsduClient(BaseOsduClient):
     def profile(self, val):
         self._profile = val
 
-    def __init__(self, data_partition_id, api_url: str = None, client_id: str = None, user: str = None, password: str = None, profile: str = None) -> None:
+    def __init__(self, data_partition_id, api_url:str=None, client_id:str=None, secret_hash:str=None,user:str=None, password:str=None, profile:str=None) -> None:
         """Authenticate and instantiate a new AWS OSDU client. Uses Cognito directly to obtain an access token.
 
         :param data_partition_id:   [Required] OSDU data partition ID, e.g. 'opendes'
@@ -28,6 +28,7 @@ class AwsOsduClient(BaseOsduClient):
                             If not provided as arg, client will attempt to load value from 
                             environment variable: OSDU_API_URL.
         :param client_id:   OSDU client ID. Must be a Cognito App Client with no client secret.
+        :param secret_hash: Amazon Cognito Secret hash. This is described here: 'https://aws.amazon.com/premiumsupport/knowledge-center/cognito-unable-to-verify-secret-hash/'
         :param user:        OSDU username. If not provided as arg, client will attempt to load value from
                             environment variable: OSDU_USER.
         :param password:    OSDU password. If not provided as arg, client will attempt to load value from
@@ -41,13 +42,14 @@ class AwsOsduClient(BaseOsduClient):
         self._client_id = client_id or os.environ.get('OSDU_CLIENT_ID')
         self._user = user or os.environ.get('OSDU_USER')
         self._profile = profile or os.environ.get('AWS_PROFILE')
+        self._secret_hash = secret_hash or os.environ.get('AWS_SECRETHASH')
         if password:
-            self.get_tokens(password)
-            password = None  # Don't leave password lying around.
+            self.get_tokens(password, secret_hash)
+            password = None # Don't leave password lying around.
         else:
-            self.get_tokens(os.environ.get('OSDU_PASSWORD'))
+            self.get_tokens(os.environ.get('OSDU_PASSWORD'), secret_hash)
 
-    def get_tokens(self, password) -> None:
+    def get_tokens(self, password, secret_hash) -> None:
         if self._profile:
             session = boto3.Session(profile_name=self._profile)
             print('Created boto3 session with profile: ', self._profile)
@@ -55,11 +57,19 @@ class AwsOsduClient(BaseOsduClient):
         else:
             cognito = boto3.client('cognito-idp')
 
+        auth_params = { 
+            'USERNAME': self._user, 
+            'PASSWORD': password
+        }
+        if secret_hash:
+            auth_params['SECRET_HASH'] = secret_hash
+        
         response = cognito.initiate_auth(
             AuthFlow='USER_PASSWORD_AUTH',
             ClientId=self._client_id,
-            AuthParameters={'USERNAME': self._user, 'PASSWORD': password}
+            AuthParameters=auth_params
         )
+
 
         self._access_token = response['AuthenticationResult']['AccessToken']
         self._refresh_token = response['AuthenticationResult']['RefreshToken']